fix(search): was not properly searching for expanded tokens in search query. bump version.
fix(search): was not expanding query words for search so a query term that was a prefix of a word in document was not searched for.
fix(updateSetAndVec): was not using expanded token for scoring document matches. This did not affect the set of returned documents but did affect the scores and could return NaN scores incorrectly.
test(search): added test case to cover expanding tokens search test case.