Publish 5.0.2
This fixes a bug in some search results returning NaN.
This adds new functions addT, removeT, searchT as variants of
of add, remove search that use a type instead of string for Error in Result.
The version of Trie updated to removes nodes when documents are
removed from the index so if you remove documents it may be
lighter on resources.